Change how nicknames are handled
Screambot used the name that was set there when no custom nickname was set. This was causing problems in servers where Screambot didn't have permission to set its nickname, now Screambot just doesn't set a nickname for itself at all if no nickname was defined for that server. I don't know why I didn't have it like this to begin with.