List servers and channels commands
New commands "servers" and "channels [serverID]" help the bot owner know what servers Screambot is a member of, and which channels in a server Screambot can scream in.

When Screambot is added to a server, he will now give way more information about the server in order to the bot owner set him up in the new server, even if the bot owner is not a member of the server themself. Yes, Screambot is a He. It's my bot and he is a He.

Also, defaults.js and embeds.js have been added to make Screambot more stable and support these new commands, respectively.

channelIDIsAllowed() has been renamed to canScreamIn(), while being changed to use has()'s code instead of its own, since they were actually just the same code.

@async tags have been removed from docstrings since I felt they were unnecessary (and I didn't even do it completely to begin with.)