Rearrange onmessage logic to fix DM behavior
One of the most recent changes added a custom response to "ayy". This did not work right in DMs since the check to see if a user had said "ayy" took place after the check if the message is in DMs (at which case, always scream), which was preventing Screambot from properly responding "lmao" when DM'd the phrase "ayy".