Fix incompatibilities with update
discord.js v12 made some breaking changes, and I didn't fix all of the incompatibilities at first that came with it when I finally switched Screambot over to v12 from v11. This commit should make Screambot perfectly compatible with discord.js v12.
The reason that is so important is because Discord is going to make breaking changes to the API in the coiming months, and discord.js will probably not update v11 of its library to adhere to these changes. For that reason, it's imperative that all my bots, including Screambot, are up-to-date with discord.js to ensure they will encounter the least trouble possible when those breaking changes hit.