Scaffold out new excitement level feature
A more intelligent way to decide when to scream. When the chat is dead, a single message should never set it off. Screambot should be more apt to scream the more activity there is in the chat.

i.e.,
randomReplyChance should be a function proportional to the number of messages recently sent in the respective channel.

Everything is set up for it now. As it stands it just returns RANDOM_REPLY_CHANCE (which may change or go away entirely). I just need to come up with a fitting mathematical formula. Off to Desmos!